+++ bot/handlers/commands/command-definitions/index.js
@@ -9,7 +9,6 @@ const {
 const {
   createRaidCheckCommandDefinition,
   createRaidSetCommandDefinition,
-  createRaidSyncCommandDefinition,
   createStatusCommandDefinition,
 } = require("./progress");
 const {
@@ -37,7 +36,6 @@ function createRaidCommandDefinitions({
     createRaidCheckCommandDefinition(),
     createRaidSetCommandDefinition(),
     createStatusCommandDefinition(),
-    createRaidSyncCommandDefinition(),
     createRaidHelpCommandDefinition(),
     createRaidGoldEarnerCommandDefinition(),
     createRemoveRosterCommandDefinition(),